requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
graphviz==0.20.1
jinja2==3.1.3
reportlab==4.1.0
python-dotenv==1.0.1
groq>=0.5.0
json-repair==0.25.0
sentence-transformers>=2.2.2
numpy>=1.26.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pdfplumber>=0.10.0
requests>=2.31.0

